Initial commit of Direction Module
Direction module is ready for testing. I made many assumptions
during development, most notably that the thresholds never need to
change and that only the channel responsible for the current detection
is relevant. My prediction (hope) is that this will work with moderate
reliability at least, and then we can further optimize it for increased
performance.